Add ZuoraClient
